backend/requirements.txt
Django==4.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
mysqlclient==2.1.1
python-decouple==3.8
openpyxl==3.1.2
pdfplumber==0.10.3
reportlab==4.0.9
ortools==9.7.2996
pandas==2.2.2
numpy==1.26.4
Pillow==10.0.0
